refactor(project): :truck: Classify files
